Add full test workflow

+++ test/par-compare.js
@@ -13,7 +13,10 @@ var exeParpar = '../bin/parpar';
 var exePar2 = 'par2';
 
 var skipFileCreate = true; // skip creating test files if they already exist (speeds up repeated failing tests, but existing files aren't checked)
+var pruneCache = false; // prune unused keys from cached results
 
+
+var fastTest = process.argv.slice(2).indexOf('-f') > -1;
 
 var fs = require('fs');
 var crypto = require('crypto');
@@ -311,7 +314,8 @@ function writeRndFile(name, size) {
 }
 writeRndFile('test64m.bin', 64*1048576);
 writeRndFile('test2200m.bin', 2200*1048576);
-writeRndFile('test4100m.bin', 4100*1048576); // >4GB to test 32-bit overflows
+if(!fastTest)
+	writeRndFile('test4100m.bin', 4100*1048576); // >4GB to test 32-bit overflows
 
 // we don't test 0 byte files - different implementations seem to treat it differently:
 // - par2cmdline: skips all 0 byte files
@@ -326,14 +330,15 @@ writeRndFile('test13m.bin', 13631477);
 
 
 var cachedResults = {};
+var setCacheKeys = {};
 var sourceFiles = {};
-
+var cacheFileName = fastTest ? 'cached-cmpref-fast.json' : 'cached-cmpref.json';
 try {
-	cachedResults = require(tmpDir + 'cached-cmpref.json');
+	cachedResults = require(tmpDir + cacheFileName);
 } catch(x) {
 	try {
 		// try current folder as well, since I tend to stick it there
-		cachedResults = require('./cached-cmpref.json');
+		cachedResults = require('./' + cacheFileName);
 	} catch(x) {
 		cachedResults = {};
 	}
@@ -456,15 +461,6 @@ var allTests = [
 		cacheKey: '22'
 	},
 	
-	// issue #6
-	{
-		in: [tmpDir + 'test64m.bin'],
-		blockSize: 40000,
-		blocks: 10000,
-		singleFile: true,
-		cacheKey: '12'
-	},
-	
 	// no recovery test
 	{
 		in: [tmpDir + 'test64m.bin'],
@@ -481,15 +477,7 @@ var allTests = [
 		cacheKey: '21'
 	},
 	
-	// 2x large block size test
-	{
-		in: [tmpDir + 'test64m.bin'],
-		blockSize: 2048*1048576 - 1024-68,
-		blocks: 1,
-		memory: is64bPlatform ? 2560*1048576 : 1536*1048576,
-		singleFile: true,
-		cacheKey: '13'
-	},
+	// large block size test
 	{
 		in: [tmpDir + 'test64m.bin'],
 		blockSize: 4294967296, // 4GB, should exceed node's limit
@@ -497,30 +485,6 @@ var allTests = [
 		memory: 511*1048576,
 		singleFile: true,
 		cacheKey: '14'
-	},
-	
-	// 2x large input file test
-	{
-		in: [tmpDir + 'test4100m.bin'],
-		blockSize: 1048576,
-		blocks: 64,
-		singleFile: true,
-		cacheKey: '15'
-	},
-	{
-		in: [tmpDir + 'test2200m.bin', tmpDir + 'test1b.bin'],
-		blockSize: 768000,
-		blocks: 2800,
-		singleFile: true,
-		cacheKey: '16'
-	},
-	
-	{ // max number of blocks test
-		in: [tmpDir + 'test64m.bin'],
-		blockSize: 2048,
-		blocks: 32768, // max allowed by par2cmdline; TODO: test w/ 65535
-		singleFile: true,
-		cacheKey: '17'
 	},
 	
 	{ // skewed slice size to test chunk miscalculation bug
@@ -531,25 +495,70 @@ var allTests = [
 		cacheKey: '18'
 	},
 	
-	{ // slice > 4GB (generally unsupported, but can be made via par2cmdline with some trickery)
-		in: [tmpDir + 'test4100m.bin'],
-		inBlocks: 1, // 4100MB slice
-		blocks: 2,
-		singleFile: true,
-		cacheKey: '19'
-	},
-	
 ];
-if(is64bPlatform) {
-	allTests.push({ // recovery > 4GB in memory [https://github.com/animetosho/par2cmdline-turbo/issues/7]
-		in: [tmpDir + 'test4100m.bin'],
-		blockSize: 100*1048576,
-		blocks: 41,
-		singleFile: true,
-		memory: 8192*1048576,
-		cacheKey: '23',
-		readSize: '100M'
-	});
+if(!fastTest) {
+	allTests.push(
+		// issue #6
+		{
+			in: [tmpDir + 'test64m.bin'],
+			blockSize: 40000,
+			blocks: 10000,
+			singleFile: true,
+			cacheKey: '12'
+		},
+		// large block+mem test
+		{
+			in: [tmpDir + 'test64m.bin'],
+			blockSize: 2048*1048576 - 1024-68,
+			blocks: 1,
+			memory: is64bPlatform ? 2560*1048576 : 1536*1048576,
+			singleFile: true,
+			cacheKey: '13'
+		},
+		
+		// 2x large input file test
+		{
+			in: [tmpDir + 'test4100m.bin'],
+			blockSize: 1048576,
+			blocks: 64,
+			singleFile: true,
+			cacheKey: '15'
+		},
+		{
+			in: [tmpDir + 'test2200m.bin', tmpDir + 'test1b.bin'],
+			blockSize: 768000,
+			blocks: 2800,
+			singleFile: true,
+			cacheKey: '16'
+		},
+		
+		{ // max number of blocks test
+			in: [tmpDir + 'test64m.bin'],
+			blockSize: 2048,
+			blocks: 32768, // max allowed by par2cmdline; TODO: test w/ 65535
+			singleFile: true,
+			cacheKey: '17'
+		},
+		
+		{ // slice > 4GB (generally unsupported, but can be made via par2cmdline with some trickery)
+			in: [tmpDir + 'test4100m.bin'],
+			inBlocks: 1, // 4100MB slice
+			blocks: 2,
+			singleFile: true,
+			cacheKey: '19'
+		},
+	);
+	if(is64bPlatform) {
+		allTests.push({ // recovery > 4GB in memory [https://github.com/animetosho/par2cmdline-turbo/issues/7]
+			in: [tmpDir + 'test4100m.bin'],
+			blockSize: 100*1048576,
+			blocks: 41,
+			singleFile: true,
+			memory: 8192*1048576,
+			cacheKey: '23',
+			readSize: '100M'
+		});
+	}
 }
 
 
@@ -635,6 +644,7 @@ async.timesSeries(allTests.length, function(testNum, cb) {
 				}
 				return ret;
 			})));
+			setCacheKeys[test.cacheKey] = 1;
 			
 			delOutput();
 			cb();
@@ -650,12 +660,19 @@ async.timesSeries(allTests.length, function(testNum, cb) {
 		fs.unlinkSync(tmpDir + 'test65k.bin');
 		fs.unlinkSync(tmpDir + 'test13m.bin');
 		fs.unlinkSync(tmpDir + 'test2200m.bin');
-		fs.unlinkSync(tmpDir + 'test4100m.bin');
+		if(!fastTest)
+			fs.unlinkSync(tmpDir + 'test4100m.bin');
 	}
 	
 	if(!err) {
+		if(pruneCache) {
+			for(var k in cachedResults)
+				if(!(k in setCacheKeys))
+					delete cachedResults[k];
+		}
+		
 		try {
-			fs.writeFileSync(tmpDir + 'cached-cmpref.json', JSON.stringify(cachedResults));
+			fs.writeFileSync(tmpDir + cacheFileName, JSON.stringify(cachedResults));
 		} catch(x) {
 			console.log(x);
 		}
